// src/types/doctor.ts
export interface Doctor {
  id: number;
  name: string;
  designation: string;
  specialties: string[];
  experience: number;
  qualification: string;
  languagesSpoken: string[];
  fees: number;
  address: string;
  consultationMode: string[];
  rating: number;
  totalReviews: number;
  about: string;
}